Fixed a bunch of compile warnings and removed an obsolete Route component that tried to use multiport, but that wont work since we need to know in which order connections were made

+++ HOPSAN++/componentLibraries/defaultLibrary/code/Hydraulic/HydraulicMultiPressureSourceC.hpp
@@ -79,7 +79,7 @@ namespace hopsan {
             mND_p_vec.resize(mNumPorts);
             mND_c_vec.resize(mNumPorts);
             mND_Zc_vec.resize(mNumPorts);
-            for (size_t i=0; i<mNumPorts; ++i)
+            for (int i=0; i<int(mNumPorts); ++i)
             {
                 mND_p_vec[i] = getSafeMultiPortNodeDataPtr(mpMP, i, NodeHydraulic::PRESSURE);
                 mND_c_vec[i] = getSafeMultiPortNodeDataPtr(mpMP, i, NodeHydraulic::WAVEVARIABLE);
